requirements.txt
discord.py==2.3.2
firebase-admin==6.2.0
python-dotenv==1.0.0
asyncio
datetime
